fix(html5/whiteboard): Debounced function recreated on every render - Remove debounced function statement from within the component. This avoids recreating the function on every render of the component, which makes the debounced function behave like a regular function (undebounced).
